Move UnitOfWorkPipelineBehavior and PublishDomainEventsPipelineBehavior to new Behaviors namespace
