+++ account-management/tests/Foundation.Tests/Application/DomainEvents/PublishDomainEventsPipelineBehaviorTests.cs
@@ -1,6 +1,7 @@
 using FluentAssertions;
 using MediatR;
 using NSubstitute;
+using PlatformPlatform.Foundation.DddCqrsFramework.Behaviors;
 using PlatformPlatform.Foundation.DddCqrsFramework.DomainEvents;
 using PlatformPlatform.Foundation.DddCqrsFramework.Persistence;
 using PlatformPlatform.Foundation.Tests.TestEntities;
