+++ account-management/tests/Foundation.Tests/Application/Persistence/UnitOfWorkPipelineBehaviorTests.cs
@@ -1,5 +1,6 @@
 using MediatR;
 using NSubstitute;
+using PlatformPlatform.Foundation.DddCqrsFramework.Behaviors;
 using PlatformPlatform.Foundation.DddCqrsFramework.Persistence;
 using PlatformPlatform.Foundation.Tests.TestEntities;
 using Xunit;
