+++ backend-foundation/src/DddCqrsFramework/Behaviors/PublishDomainEventsPipelineBehavior.cs
@@ -1,7 +1,8 @@
 using MediatR;
+using PlatformPlatform.Foundation.DddCqrsFramework.DomainEvents;
 using PlatformPlatform.Foundation.DddCqrsFramework.Persistence;
 
-namespace PlatformPlatform.Foundation.DddCqrsFramework.DomainEvents;
+namespace PlatformPlatform.Foundation.DddCqrsFramework.Behaviors;
 
 /// <summary>
 ///     This method publishes any domain events that were generated during the execution of a command (and added to
