+++ backend-foundation/src/DddCqrsFramework/DddCqrsFramework.csproj
@@ -9,9 +9,9 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="IdGen" Version="3.0.3"/>
-        <PackageReference Include="JetBrains.Annotations" Version="2022.3.1"/>
-        <PackageReference Include="MediatR" Version="12.0.1"/>
+        <PackageReference Include="IdGen" Version="3.0.3" />
+        <PackageReference Include="JetBrains.Annotations" Version="2022.3.1" />
+        <PackageReference Include="MediatR" Version="12.0.1" />
     </ItemGroup>
 
 </Project>
